Create the code (there's... a lot)
the prompt given was 'hi. please make me a web forum based on html called 'science stuff'. there should be three main repositories called biology, chemistry, and physics. make biology colour #d1f2d0, make chemistry colour #abeaff, and make physics colour #ffb5b5. each of these will be the background colour of the entire repository, and the text will be in #000000. make ranks such as admin/op, (administrator/operator), member, viewer and founder as well as a sign in and sign up system with username or email, and password. this can be linked to google, and users can sign in from a topic page, be transported to the sign in page (which is opened on a new tab), and then transported back to the topic page they were viewing, with the sign in page closing when it is no longer required. make it so that people can view, post and share things such as their own repositories. however, all large repositories (general topics) must be approved by an admin/op or ME (the founder) and all other posts will be approved by a safeguarding ai. the posts can be new repositories, forks from a repository, a question regarding a topic (a comment), an animation to explain a topic, (animations can be made with desmos which would be linked, or with shapes and manual rotation and dragging which is then recorded as an animation). Viewers cannot post, but can only comment and this is a temporary sanction for those who try to do anything purposefully with ill intent. each user has 3 chances to do something with ill intent. once these are gone, they are banned along with their email and device. if this appears to be a mistake, the user in question may send an email to mewbenedictarnoldzilla@gmail.com, and I will personally check if their request is valid. Make it so that when halfway through an animation, people can save it to their account and save it there (offline use permitted) before sharing/sending it with a button in the top right hand corner called ‘share/send’. Make it so that in the animation making page, there is an option to add a polygon of n sides, and the users can set n to anything they want, and that each connection between two lines is represented as an unfilled but outlined dot, of an adjustable colour (adjusting this colour will be in settings), and this connection can be cut or created when a dot is dragged to another one. Connect the website to claude sonnet 4, allowing them to also get it to do the work for them with it being an ‘ai assistant’ appearing as a small box on the chat page, with it being a minimisable popup page. Next, make it so that people can make repositories with a button in the top right hand corner of their screen, which takes them to a page with an option for an image, a title, and a description, as well as if it is part of biology, chemistry, or physics which will determine the background colour, although the text can be changed to anything except the background colour. If it is too hard to read, complaints can also be sent to mewbenedictarnoldzilla@gmail.com, with there being a button for ‘file a complaint to mewbenedictarnoldzilla@gmail.com’ and a smaller button underneath which says ‘copy email address’. Titles and headings and subheadings will have larger text than normal text'.



rather detailed.